Compiles and links when helpers.h/.cc are is split to two files
However, including the .cc file in FunctionLibary.h is still necessary
because I can't figure out how to get it to link properly.
